requirements.txt
hug
flask
numpy
textblob
gunicorn
http://download.pytorch.org/whl/cpu/torch-0.3.1-cp35-cp35m-linux_x86_64.whl 
torchvision
